+++ foxes/models/turbine_types/PCt_single_files.py
@@ -7,18 +7,21 @@ from foxes.utils import PandasFileHelper
 from foxes.data import PCTCURVE, parse_Pct_file_name
 import foxes.variables as FV
 
-
-class PCtFile(TurbineType):
+class PCtSingleFiles(TurbineType):
     """
     Calculate power and ct by interpolating
-    from power-ct-curve data file.
+    from power-curve and ct-curve data files.
 
     Parameters
     ----------
-    data_source : str or pandas.DataFrame
-        The file path, static name, or data
-    col_ws : str
-        The wind speed column
+    data_source_P : str or pandas.DataFrame
+        The file path for the power-curve, static name, or data
+    data_source_ct : str or pandas.DataFrame
+        The file path for the ct-curve, static name, or data 
+    col_ws_P_file : str
+        The wind speed column in the file of the power-curve
+    col_ws_ct_file : str
+        The wind speed column in the file of the ct-curve
     col_P : str
         The power column
     col_ct : str
@@ -36,13 +39,19 @@ class PCtFile(TurbineType):
         The wind speed variable for ct lookup
     var_ws_P : str
         The wind speed variable for power lookup
+    pd_file_read_pars_P:  dict, optional
+        Parameters for pandas power file reading
+    pd_file_read_pars_ct:  dict, optional
+        Parameters for pandas ct file reading
     **parameters : dict, optional
         Parameters for pandas file reading
 
     Attributes
     ----------
-    source : str or pandas.DataFrame
-        The file path, static name, or data
+    source_P : str or pandas.DataFrame
+        The file path for the power-curve, static name, or data
+    source_ct : str or pandas.DataFrame
+        The file path for the ct-curve, static name, or data
     col_ws : str
         The wind speed column
     col_P : str
@@ -58,45 +67,48 @@ class PCtFile(TurbineType):
         The wind speed variable for ct lookup
     WSP : str
         The wind speed variable for power lookup
-    rpars : dict, optional
-        Parameters for pandas file reading
+    rpars_P : dict, optional
+        Parameters for pandas power file reading
+    rpars_ct : dict, optional
+        Parameters for pandas ct file reading
 
     """
 
     def __init__(
         self,
-        data_source,
-        col_ws="ws",
+        data_source_P,
+        data_source_ct,
+        col_ws_P_file="ws",
+        col_ws_ct_file="ws",
         col_P="P",
         col_ct="ct",
         rho=None,
         flag_yawm=False,
-        p_ct=1.0,
+        p_ct=1.,
         p_P=1.88,
         var_ws_ct=FV.REWS2,
         var_ws_P=FV.REWS3,
-        pd_file_read_pars={},
+        pd_file_read_pars_P={},
+        pd_file_read_pars_ct={},
         **parameters
     ):
-        if not isinstance(data_source, pd.DataFrame):
-            pars = parse_Pct_file_name(data_source)
-            pars.update(parameters)
-        else:
-            pars = parameters
-
+        pars = parameters # no parsing because two files are given
         super().__init__(**pars)
 
-        self.source = data_source
-        self.col_ws = col_ws
-        self.col_P = col_P
-        self.col_ct = col_ct
-        self.rho = rho
-        self.flag_yawm = flag_yawm
-        self.p_ct = p_ct
-        self.p_P = p_P
-        self.WSCT = var_ws_ct
-        self.WSP = var_ws_P
-        self.rpars = pd_file_read_pars
+        self.source_P       = data_source_P
+        self.source_ct      = data_source_ct
+        self.col_ws_P_file  = col_ws_P_file
+        self.col_ws_ct_file = col_ws_ct_file
+        self.col_P          = col_P
+        self.col_ct         = col_ct
+        self.rho            = rho
+        self.flag_yawm      = flag_yawm
+        self.p_ct           = p_ct
+        self.p_P            = p_P
+        self.WSCT           = var_ws_ct
+        self.WSP            = var_ws_P
+        self.rpars_P        = pd_file_read_pars_P
+        self.rpars_ct       = pd_file_read_pars_ct
 
     def output_farm_vars(self, algo):
         """
@@ -106,7 +118,7 @@ class PCtFile(TurbineType):
         ----------
         algo : foxes.core.Algorithm
             The calculation algorithm
-
+        
         Returns
         -------
         output_vars : list of str
@@ -114,7 +126,7 @@ class PCtFile(TurbineType):
 
         """
         return [FV.P, FV.CT]
-
+    
     def initialize(self, algo, st_sel, verbosity=0):
         """
         Initializes the model.
@@ -130,20 +142,30 @@ class PCtFile(TurbineType):
             The verbosity level
 
         """
-        if isinstance(self.source, pd.DataFrame):
-            data = self.source
+        #read power-curve
+        if isinstance(self.source_P, pd.DataFrame):
+            data_P  = self.source_P
         else:
-            fpath = algo.dbook.get_file_path(PCTCURVE, self.source, check_raw=True)
-            data = PandasFileHelper.read_file(fpath, **self.rpars)
-
-        data = data.set_index(self.col_ws).sort_index()
-        self.data_ws = data.index.to_numpy()
-        self.data_P = data[self.col_P].to_numpy()
-        self.data_ct = data[self.col_ct].to_numpy()
+            fpath   = algo.dbook.get_file_path(PCTCURVE, self.source_P, check_raw=True)
+            data_P  = PandasFileHelper.read_file(fpath, **self.rpars_P)
+            
+        #read ct-curve
+        if isinstance(self.source_ct, pd.DataFrame):
+            data_ct  = self.source_ct
+        else:
+            fpath   = algo.dbook.get_file_path(PCTCURVE, self.source_ct, check_raw=True)
+            data_ct  = PandasFileHelper.read_file(fpath, **self.rpars_ct)
+            
+        data_P          = data_P.set_index(self.col_ws_P_file).sort_index()
+        data_ct         = data_ct.set_index(self.col_ws_ct_file).sort_index()
+        self.data_ws_P  = data_P.index.to_numpy()
+        self.data_ws_ct = data_ct.index.to_numpy()
+        self.data_P     = data_P[self.col_P].to_numpy()
+        self.data_ct    = data_ct[self.col_ct].to_numpy()
         super().initialize(algo, st_sel, verbosity=verbosity)
-
+    
     def calculate(self, algo, mdata, fdata, st_sel):
-        """ "
+        """"
         The main model calculation.
 
         This function is executed on a single chunk of data,
@@ -160,26 +182,26 @@ class PCtFile(TurbineType):
         st_sel : numpy.ndarray of bool
             The state-turbine selection,
             shape: (n_states, n_turbines)
-
+            
         Returns
         -------
         results : dict
             The resulting data, keys: output variable str.
             Values: numpy.ndarray with shape (n_states, n_turbines)
 
-        """
+        """  
         rews2 = fdata[self.WSCT][st_sel]
         rews3 = fdata[self.WSP][st_sel]
 
         # apply air density correction:
         if self.rho is not None:
-
+            
             # correct wind speed by air density, such
             # that in the partial load region the
             # correct value is reconstructed:
-            rho = fdata[FV.RHO][st_sel]
-            rews2 *= (self.rho / rho) ** 0.5
-            rews3 *= (self.rho / rho) ** (1.0 / 3.0)
+            rho    = fdata[FV.RHO][st_sel]
+            rews2 *= ( self.rho / rho )**0.5
+            rews3 *= ( self.rho / rho )**(1./3.) 
             del rho
 
         # in yawed case, calc yaw corrected wind speed:
@@ -188,26 +210,22 @@ class PCtFile(TurbineType):
             # calculate corrected wind speed wsc,
             # gives ws**3 * cos**p_P in partial load region
             # and smoothly deals with full load region:
-            yawm = fdata[FV.YAWM][st_sel]
-            cosm = np.cos(yawm / 180 * np.pi)
-            rews2 *= (cosm**self.p_ct) ** 0.5
-            rews3 *= (cosm**self.p_P) ** (1.0 / 3.0)
+            yawm   = fdata[FV.YAWM][st_sel]
+            cosm   = np.cos(yawm / 180 * np.pi)
+            rews2 *= ( cosm**self.p_ct )**0.5
+            rews3 *= ( cosm**self.p_P )**(1./3.) 
             del yawm, cosm
 
         out = {
-            FV.P: fdata.get(FV.P, np.zeros_like(fdata[self.WSCT])),
-            FV.CT: fdata.get(FV.CT, np.zeros_like(fdata[self.WSP])),
+            FV.P : fdata.get(FV.P, np.zeros_like(fdata[self.WSCT])),
+            FV.CT: fdata.get(FV.CT, np.zeros_like(fdata[self.WSP]))
         }
 
-        out[FV.P][st_sel] = np.interp(
-            rews3, self.data_ws, self.data_P, left=0.0, right=0.0
-        )
-        out[FV.CT][st_sel] = np.interp(
-            rews2, self.data_ws, self.data_ct, left=0.0, right=0.0
-        )
+        out[FV.P][st_sel]  = np.interp(rews3, self.data_ws_P, self.data_P, left=0., right=0.)
+        out[FV.CT][st_sel] = np.interp(rews2, self.data_ws_ct, self.data_ct, left=0., right=0.)
 
         return out
-
+    
     def finalize(self, algo, results, st_sel, clear_mem=False, verbosity=0):
         """
         Finalizes the model.
@@ -229,5 +247,6 @@ class PCtFile(TurbineType):
 
         """
         if clear_mem:
-            del self.data_ws, self.data_P, self.data_ct
+            del self.data_ws_P, self.data_ws_ct, self.data_P, self.data_ct
         super().finalize(algo, results, clear_mem, verbosity=verbosity)
+