+++ foxes/models/vertical_profiles/abl_log/abl_log_unstable_ws.py
@@ -1,5 +1,5 @@
 from foxes.core import VerticalProfile
-from foxes.tools.abl import unstable
+from foxes.utils.abl import unstable
 import foxes.constants as FC
 import foxes.variables as FV