+++ foxes/models/wake_models/wind/__init__.py
@@ -1,0 +1,3 @@
+from .jensen import JensenWake
+from .bastankhah import BastankhahWake
+from .turbopark import TurbOParkWake