+++ docs/source/history.md
@@ -34,4 +34,9 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
 ## v0.1.4-alpha
 
 - Fixes
-    - Static data: adding missing data `wind_rotation.nc` to manifest
+    - Static data: Adding missing data `wind_rotation.nc` to manifest
+- Models
+    - New wake model added: `TurbOParkWake` from Orsted
+    - New turbine type added: `PCtSingleFiles`, reads power and thrust curves from two separate files
+
+**Full Changelog**: https://github.com/FraunhoferIWES/foxes/commits/v0.1.4